Refactor code to use date-fns for date formatting and relative time
Updated the applogic.js and domModule.js files to utilize the date-fns library for formatting and displaying dates. Imported the necessary functions from date-fns and modified the code to use them. Updated the getDueDate function in applogic.js to use the format function for date formatting. In domModule.js, used the formatDistanceToNow function to display relative time for due dates. Also, updated import statements and added necessary dependencies. This improves the consistency and accuracy of date handling in the app.